formatter run

+++ contracts/interfaces/iTracker.sol
@@ -3,6 +3,15 @@ pragma solidity ^0.8.9;
 /* SPDX-License-Identifier: UNLICENSED */
 
 interface iTracker {
-    function processDeposit(address, uint, uint) external;
-    function processWithdraw(address, uint, uint) external;
+    function processDeposit(
+        address,
+        uint256,
+        uint256
+    ) external;
+
+    function processWithdraw(
+        address,
+        uint256,
+        uint256
+    ) external;
 }
